requirements.txt
beautifulsoup4
matplotlib
textblob
flask
gunicorn
pytesseract
gingerit
spacy
pillow
nltk
requests
googlemaps
pandas
selenium
lxml